+++ tests/Microsoft.Identity.Test.Integration.netcore/Microsoft.Identity.Test.Integration.NetCore.csproj
@@ -10,6 +10,7 @@
 
     <Compile Include="../Microsoft.Identity.Test.Integration.NetFx/**/*.cs" Exclude="../Microsoft.Identity.Test.Integration.NetFx/obj/**/*" />
     <None Include="../Microsoft.Identity.Test.Integration.NetFx/federationMetadata.xml" />
+    <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client.Brokers\Microsoft.Identity.Client.Broker.csproj" />
     
     <ProjectReference Include="..\..\src\client\Microsoft.Identity.Client\Microsoft.Identity.Client.csproj" />
     <ProjectReference Include="..\Microsoft.Identity.Test.Core.UIAutomation\Microsoft.Identity.Test.UIAutomation.Infrastructure.csproj" />
